Introduce a separate FilePropertyVisitor
that is generic and not bound to tasks.
