+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/SnapshotTransformInputsBuildOperationResult.java
@@ -17,15 +17,158 @@
 package org.gradle.api.internal.artifacts.transform;
 
 import org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult;
+import org.gradle.api.internal.tasks.FilePropertyVisitState;
 import org.gradle.api.internal.tasks.properties.InputFilePropertySpec;
 import org.gradle.internal.execution.caching.CachingState;
+import org.gradle.internal.execution.history.BeforeExecutionState;
+import org.gradle.internal.hash.HashCode;
 import org.gradle.operations.dependencies.transforms.SnapshotTransformInputsBuildOperationType;
+import org.gradle.operations.execution.FilePropertyVisitor;
 
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Deque;
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.TreeMap;
 
 public class SnapshotTransformInputsBuildOperationResult extends BaseSnapshotInputsBuildOperationResult implements SnapshotTransformInputsBuildOperationType.Result {
 
-    public SnapshotTransformInputsBuildOperationResult(CachingState cachingState, Set<InputFilePropertySpec> inputFileProperties) {
-        super(cachingState, inputFileProperties);
+    private final Set<InputFilePropertySpec> inputFilePropertySpecs;
+
+    public SnapshotTransformInputsBuildOperationResult(CachingState cachingState, Set<InputFilePropertySpec> inputFilePropertySpecs) {
+        super(cachingState);
+        this.inputFilePropertySpecs = inputFilePropertySpecs;
+    }
+
+    @Override
+    public void visitInputFileProperties(FilePropertyVisitor visitor) {
+        getBeforeExecutionState()
+            .map(BeforeExecutionState::getInputFileProperties)
+            .ifPresent(inputFileProperties -> FilePropertyVisitState.visitInputFileProperties(inputFileProperties, visitor, inputFilePropertySpecs));
+    }
+
+    @Override
+    protected Map<String, Object> fileProperties() {
+        final Map<String, Object> fileProperties = new TreeMap<>();
+        visitInputFileProperties(new FilePropertyVisitor() {
+            Property property;
+            final Deque<DirEntry> dirStack = new ArrayDeque<>();
+
+            class Property {
+                private final String hash;
+                private final Set<String> attributes;
+                private final List<Entry> roots = new ArrayList<>();
+
+                public Property(String hash, Set<String> attributes) {
+                    this.hash = hash;
+                    this.attributes = attributes;
+                }
+
+                public String getHash() {
+                    return hash;
+                }
+
+                public Set<String> getAttributes() {
+                    return attributes;
+                }
+
+                public Collection<Entry> getRoots() {
+                    return roots;
+                }
+            }
+
+            abstract class Entry {
+                private final String path;
+
+                public Entry(String path) {
+                    this.path = path;
+                }
+
+                public String getPath() {
+                    return path;
+                }
+
+            }
+
+            class FileEntry extends Entry {
+                private final String hash;
+
+                FileEntry(String path, String hash) {
+                    super(path);
+                    this.hash = hash;
+                }
+
+                public String getHash() {
+                    return hash;
+                }
+            }
+
+            class DirEntry extends Entry {
+                private final List<Entry> children = new ArrayList<>();
+
+                DirEntry(String path) {
+                    super(path);
+                }
+
+                public Collection<Entry> getChildren() {
+                    return children;
+                }
+            }
+
+            @Override
+            public void preProperty(VisitState state) {
+                property = new Property(HashCode.fromBytes(state.getPropertyHashBytes()).toString(), state.getPropertyAttributes());
+                fileProperties.put(state.getPropertyName(), property);
+            }
+
+            @Override
+            public void preRoot(VisitState state) {
+
+            }
+
+            @Override
+            public void preDirectory(VisitState state) {
+                boolean isRoot = dirStack.isEmpty();
+                DirEntry dir = new DirEntry(isRoot ? state.getPath() : state.getName());
+                if (isRoot) {
+                    property.roots.add(dir);
+                } else {
+                    //noinspection ConstantConditions
+                    dirStack.peek().children.add(dir);
+                }
+                dirStack.push(dir);
+            }
+
+            @Override
+            public void file(VisitState state) {
+                boolean isRoot = dirStack.isEmpty();
+                FileEntry file = new FileEntry(isRoot ? state.getPath() : state.getName(), HashCode.fromBytes(state.getHashBytes()).toString());
+                if (isRoot) {
+                    property.roots.add(file);
+                } else {
+                    //noinspection ConstantConditions
+                    dirStack.peek().children.add(file);
+                }
+            }
+
+            @Override
+            public void postDirectory() {
+                dirStack.pop();
+            }
+
+            @Override
+            public void postRoot() {
+
+            }
+
+            @Override
+            public void postProperty() {
+
+            }
+        });
+        return fileProperties;
     }
 }
